requirements.txt
ultralytics
roboflow
python-dotenv
opencv-python